Changes to css and login
Changed the colors in the layout. Also fixed the login so that it redirects the user to the school profile html (if the user has entered a school). If the school that the user entered in the school field does not exist in the school database, it displays a not found message in the school profile html.