Fixed tables.py and default.py, added a review html
Mostly fixed some of the fields in the tables and the methods in default.py. Also added a review html that can be accessed by hovering over the class name (which I made into a hyperlink). Also made the add school, add class, and add review links accessible only to logged-in users.